Addressed two 64 bit compiler warnings.
  * misc/openvas_server.c (openvas_server_send): Use %zu for size_t.

  * omp/xml.c (read_entity_and_string): Cast to match format string.


git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@9421 423fd1db-d629-0410-8442-d21db03e70f4

+++ omp/xml.c
@@ -451,7 +451,7 @@ read_entity_and_string (gnutls_session_t * session, entity_t * entity,
       ssize_t count;
       while (1)
         {
-          g_message ("   asking for %i\n", buffer_end - buffer_start);
+          g_message ("   asking for %i\n", (int) (buffer_end - buffer_start));
           count =
             gnutls_record_recv (*session, buffer_start,
                                 buffer_end - buffer_start);
